Bug with attributes should be double not int.
added web version first part completed for loading patient attributes and getting the QG ID.. next phase is to create the question controls and page

+++ AMDES_KBS/AMDES_KBS/Controllers/CLIPSController.cs
@@ -171,7 +171,7 @@ namespace AMDES_KBS.Controllers
         {
             assert(new StringBuilder("(attribute AGE " + CurrentPatient.getAge() + ")"), false);
 
-            foreach (KeyValuePair<string, int> kvp in CurrentPatient.getAttributes())
+            foreach (KeyValuePair<string, double> kvp in CurrentPatient.getAttributes())
             {
                 assert(new StringBuilder("(attribute " + kvp.Key.ToUpper().Replace(" ", "_") + " " + kvp.Value.ToString() + ")"), false);
             }
